Move to utilities.go
