+++ pkg/codegen/nodejs/gen.go
@@ -64,17 +64,6 @@ func title(s string) string {
 	return string(append([]rune{unicode.ToUpper(runes[0])}, runes[1:]...))
 }
 
-// escape returns the string escaped for a JS string literal
-func escape(s string) string {
-	// Seems the most fool-proof way of doing this is by using the JSON marshaler and then stripping the surrounding quotes
-	escaped, err := json.Marshal(s)
-	contract.AssertNoError(err)
-	contract.Assertf(len(escaped) >= 2, "JSON(%s) expected a quoted string but returned %s", s, escaped)
-	contract.Assertf(escaped[0] == byte('"') && escaped[len(escaped)-1] == byte('"'), "JSON(%s) expected a quoted string but returned %s", s, escaped)
-
-	return string(escaped)[1:(len(escaped) - 1)]
-}
-
 // camel converts s to camel case.
 //
 // Examples:
